refactor(backend): eval 패키지 계층 분리(domain/repository) 및 참조 정리\n\n- domain/* 로 엔티티 이동\n- repository/* 로 레포 분리\n- 서비스/컨트롤러/리포트 의존 경로 수정\n- 로컬 빌드 성공
