+++ backend/src/main/java/com/company/evaluation/eval/EvalController.java
@@ -1,6 +1,9 @@
 package com.company.evaluation.eval;
 
 import com.company.evaluation.common.ApiResponse;
+import com.company.evaluation.eval.domain.EmployeeMemo;
+import com.company.evaluation.eval.domain.Evaluation;
+import com.company.evaluation.eval.domain.EvalCategory;
 import com.company.evaluation.eval.dto.EvalItemRequest;
 import com.company.evaluation.eval.dto.EvalItemResponse;
 import jakarta.validation.Valid;
