+++ backend/src/main/java/com/company/evaluation/eval/EvalService.java
@@ -2,6 +2,8 @@ package com.company.evaluation.eval;
 
 import com.company.evaluation.employee.Employee;
 import com.company.evaluation.employee.EmployeeRepository;
+import com.company.evaluation.eval.domain.*;
+import com.company.evaluation.eval.repository.*;
 import com.company.evaluation.eval.dto.EvalItemRequest;
 import com.company.evaluation.eval.dto.EvalItemResponse;
 import com.company.evaluation.eval.dto.EvaluationRowResponse;
