+++ backend/src/main/java/com/company/evaluation/eval/SalaryRaiseDto.java
@@ -3,6 +3,7 @@ package com.company.evaluation.eval;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import com.company.evaluation.eval.domain.SalaryRaise;
 
 @Data
 @NoArgsConstructor
