+++ backend/src/main/java/com/company/evaluation/eval/repository/EvalCategoryRepository.java
@@ -1,5 +1,6 @@
-package com.company.evaluation.eval;
+package com.company.evaluation.eval.repository;
 
+import com.company.evaluation.eval.domain.EvalCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EvalCategoryRepository extends JpaRepository<EvalCategory, Long> {}
