+++ backend/src/main/java/com/company/evaluation/eval/repository/EvalItemRepository.java
@@ -1,5 +1,6 @@
-package com.company.evaluation.eval;
+package com.company.evaluation.eval.repository;
 
+import com.company.evaluation.eval.domain.EvalItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EvalItemRepository extends JpaRepository<EvalItem, Long> {}
