+++ backend/src/main/java/com/company/evaluation/eval/repository/SalaryRaiseRepository.java
@@ -1,5 +1,6 @@
-package com.company.evaluation.eval;
+package com.company.evaluation.eval.repository;
 
+import com.company.evaluation.eval.domain.SalaryRaise;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
